Made norton script watch friendly
- watch -n1 killnorton.sh will kill the stupid process every 1 sec.

+++ bin/killnorton.sh
@@ -2,4 +2,9 @@
 
 pid=`ps ax | grep SymAutoProtect | grep -v grep | awk '{ print $1 }'`
 
-sudo kill $pid
+if [[ -z "$pid" ]]; then
+   echo "not running"
+else
+   echo "running as $pid, killing..."
+   sudo kill $pid
+fi
